+++ winter-common/src/main/java/com/winter/common/core/page/TableDataInfo.java
@@ -24,7 +24,7 @@ public class TableDataInfo implements Serializable {
      * 总记录数
      */
     @ApiModelProperty(value = "总记录数")
-    private int total;
+    private long total;
 
     /**
      * 列表数据
@@ -56,7 +56,7 @@ public class TableDataInfo implements Serializable {
      * @param list  列表数据
      * @param total 总记录数
      */
-    public TableDataInfo(List<?> list, int total) {
+    public TableDataInfo(List<?> list, long total) {
         this.rows = list;
         this.total = total;
     }
